gameplayservice/gameplay/requirements.txt
djangorestframework
requests
python-dotenv
djangorestframework-simplejwt
psycopg[binary,pool]
gunicorn
python-logstash